requirements.txt
annotated-types==0.6.0
anyio==4.0.0
attrs==23.2.0
blis==0.7.11
build==1.0.3
catalogue==2.0.10
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
distro==1.8.0
dnspython==2.4.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fastapi==0.100.0
flake8==6.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.1
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
iniconfig==2.0.0
ipaddress==1.0.23
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
motor==3.3.2
murmurhash==1.0.10
numpy>=1.24.0
openai==1.6.1
outcome==1.3.0.post0
packaging==23.2
pandas>=2.0.0
pip-tools==7.3.0
pluggy==1.3.0
preshed==3.0.9
priority==2.0.0
pycodestyle==2.11.1
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.1.0
pymongo==4.6.1
pyproject_hooks==1.0.0
PySocks==1.7.1
pytest==7.4.3
pytest-asyncio==0.23.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.16.0
setuptools==68.2.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.27.0
thinc==8.2.2
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.25.0
wasabi==1.1.2
weasel==0.3.4
websockets==12.0
wheel==0.41.3
wsproto==1.2.0
